Update FinalsLFP.jsx and ProbleShower.jsx to add support for displaying Exponential Functions problems
The code changes in FinalsLFP.jsx and ProbleShower.jsx enable the display of problems related to Exponential Functions on the Finals page. The FinalsLFP component is updated to render the Exponential Functions problems when the topic parameter is set to 'Exponential-Functions'. The necessary imports and routing changes have been made to support this feature.